requirements.txt
python-fasthtml
shad4fast

